services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.7
langchain-openai==0.1.8
openai==1.35.10
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
tortoise-orm==0.20.0
asyncpg==0.29.0
aiofiles==23.2.1
sse-starlette==1.8.2
structlog==24.1.0
celery==5.3.6
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-logging==0.46b0
prometheus-fastapi-instrumentator==6.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.4
langchain-community==0.2.7
pypdf==4.2.0
python-docx==1.1.2
rag-shared @ file:///app/packages/rag_shared
